(PC-12501) Move educonnnect connector to connector repository
