+++ api/src/pcapi/connectors/beneficiaries/educonnect/exceptions.py
